added a state managmenet context for main dialob admin view component
