+++ frontend/dialob-dashboard-material/src/DialobAdmin.tsx
@@ -11,8 +11,9 @@ import {
   enGB as enLocale,
   ms as msLocale
 } from 'date-fns/locale';
-import { DialobAdminViewProps } from "./types";
+import { DialobAdminProps } from "./types";
 import { DialobDashboardFetchProvider } from "./context";
+import { DialobDashboardStateProvider } from "./context/DialobDashboardStateContext";
 
 const localeMap: { [key: string]: any } = {
   en: enLocale,
@@ -22,12 +23,14 @@ const localeMap: { [key: string]: any } = {
   ms: msLocale,
 };
 
-export const DialobAdmin: React.FC<DialobAdminViewProps> = ({ config, showNotification }) => {
+export const DialobAdmin: React.FC<DialobAdminProps> = ({ config, showNotification }) => {
   return (
     <DialobDashboardFetchProvider>
       <LocalizationProvider dateAdapter={AdapterDateFns} adapterLocale={localeMap[config.language]}>
         <IntlProvider locale={config.language || 'en'} messages={messages[config.language]}>
-          <DialobAdminView config={config} showNotification={showNotification} />
+          <DialobDashboardStateProvider config={config} showNotification={showNotification}>
+            <DialobAdminView />
+          </DialobDashboardStateProvider>
         </IntlProvider>
       </LocalizationProvider>
     </DialobDashboardFetchProvider>
