+++ frontend/dialob-dashboard-material/src/context/index.ts
@@ -1,1 +1,2 @@
-export * from './fetchContext';
+export * from './FetchContext';
+export * from './DialobDashboardStateContext';
