+++ frontend/dialob-dashboard-material/src/index.tsx
@@ -1,11 +1,12 @@
 /* eslint-disable react-refresh/only-export-components */
-export * from './components'
-
 export { DialobAdmin } from './DialobAdmin';
 export { DialobAdminView } from './DialobAdminView';
 
-export * from './context';
-
 export { messages } from './intl';
 
+export * from './components'
+export * from './context';
+export * from './backend';
+export * from './util';
 export * from './types';
+export * from './middleware';
